@@ -1,13 +1,35 @@
-module github.com/siddontang/go-mysql-elasticsearch
+module github.com/go-mysql-org/go-mysql-elasticsearch
 
-go 1.12
+go 1.18
 
 require (
-	github.com/BurntSushi/toml v0.3.1
-	github.com/juju/errors v0.0.0-20190207033735-e65537c515d7
+	github.com/BurntSushi/toml v1.3.2
+	github.com/go-mysql-org/go-mysql v1.7.0
+	github.com/juju/errors v1.0.0
 	github.com/pingcap/check v0.0.0-20190102082844-67f458068fc8
-	github.com/prometheus/client_golang v0.9.3
+	github.com/prometheus/client_golang v1.17.0
 	github.com/siddontang/go v0.0.0-20180604090527-bdc77568d726
 	github.com/siddontang/go-log v0.0.0-20190221022429-1e957dd83bed
-	github.com/siddontang/go-mysql v0.0.0-20190524062908-de6c3a84bcbe
+)
+
+require (
+	github.com/beorn7/perks v1.0.1 // indirect
+	github.com/cespare/xxhash/v2 v2.2.0 // indirect
+	github.com/golang/protobuf v1.5.3 // indirect
+	github.com/google/uuid v1.3.0 // indirect
+	github.com/matttproud/golang_protobuf_extensions v1.0.4 // indirect
+	github.com/pingcap/errors v0.11.5-0.20210425183316-da1aaba5fb63 // indirect
+	github.com/pingcap/log v0.0.0-20210625125904-98ed8e2eb1c7 // indirect
+	github.com/pingcap/tidb/parser v0.0.0-20221126021158-6b02a5d8ba7d // indirect
+	github.com/prometheus/client_model v0.4.1-0.20230718164431-9a2bf3000d16 // indirect
+	github.com/prometheus/common v0.44.0 // indirect
+	github.com/prometheus/procfs v0.11.1 // indirect
+	github.com/shopspring/decimal v0.0.0-20180709203117-cd690d0c9e24 // indirect
+	go.uber.org/atomic v1.7.0 // indirect
+	go.uber.org/multierr v1.6.0 // indirect
+	go.uber.org/zap v1.18.1 // indirect
+	golang.org/x/sys v0.11.0 // indirect
+	golang.org/x/text v0.9.0 // indirect
+	google.golang.org/protobuf v1.31.0 // indirect
+	gopkg.in/natefinch/lumberjack.v2 v2.0.0 // indirect
 )
