@@ -7,10 +7,10 @@ import (
 	"strings"
 	"sync"
 
+	"github.com/go-mysql-org/go-mysql-elasticsearch/elastic"
+	"github.com/go-mysql-org/go-mysql/canal"
 	"github.com/juju/errors"
 	"github.com/siddontang/go-log/log"
-	"github.com/siddontang/go-mysql-elasticsearch/elastic"
-	"github.com/siddontang/go-mysql/canal"
 )
 
 // ErrRuleNotExist is the error if rule is not defined.
